galmark/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "galmark"
version = "0.1.0"

dependencies = [
  "astropy==6.1.3",
  "pillow==10.4.0",
  "PyQt6==6.7.1",
  "pandas==2.2.2"
]

authors = [
  {name = "Andi Kisare", email = "akisare@outlook.com"},
  {name = "Ryan Walker", email = "astrorya@umich.edu"}
]

[project.scripts]
galmark = "galmark:main"
